+ 12-bit PWM back to 8-bit

+++ src/pwm_audio.h
@@ -4,7 +4,7 @@ void init_sound();
 void pcm_setup(int hz, size_t size);
 void pcm_cleanup(void);
 typedef volatile int16_t* (*pcm_end_callback_t)(volatile size_t* size);
-void pcm_set_buffer(int16_t* buff, uint8_t channels, size_t size, pcm_end_callback_t cb);
+void pcm_set_buffer(uint8_t* buff, uint8_t channels, size_t size, pcm_end_callback_t cb);
 // internal call on core#1
 void pcm_call();
 bool pcm_data_in(void);
